move stream beacon test to separate file
